requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.0
httpx<0.25.0,>=0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==24.1.0